tests/wpt/web-platform-tests/html/dom/dynamic-markup-insertion/document-write/005.html

<!doctype html>
<title>document.write</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
test(
function() {
  document.write("<i>");
  document.write("Filler Text");
  assert_equals(document.body.firstChild.localName, "i");
  assert_equals(document.body.firstChild.textContent, "Filler Text");
}
);
</script>
<div id="log"></div>
